# openapi.json
{"components":{"securitySchemes":{"providerKey":{"description":"Your own AI provider API key (e.g. OpenRouter). Muse stores it sealed in its vault and sends it with each call; the server uses it only for that one Pi run and never stores it.","in":"header","name":"X-Pi-Api-Key","type":"apiKey"}}},"info":{"contact":{"name":"Pi Agent","url":"https://github.com/earendil-works/pi"},"description":"Ask Pi agent (https://github.com/earendil-works/pi), the terminal coding agent, to do coding tasks. Bring your own provider API key: it is sent per call, used for that one Pi run only, and never stored. Submit a task, get a job id back immediately, then poll for the result. Jobs run one at a time in a fresh sandbox with a hard timeout.","title":"Pi - Muse Connector API","version":"1.1.0"},"openapi":"3.0.3","paths":{"/api/pi-task":{"post":{"description":"Submit a natural-language coding task to Pi (write a script, fix a bug, explain code, scaffold a project...). Send your provider API key in the X-Pi-Api-Key header. Returns 202 with a job_id immediately — the task runs asynchronously. Poll GET /api/pi-task/{job_id} until status is done, failed, or timeout. Rate limited per IP.","parameters":[{"description":"Provider name (default: server default, usually openrouter).","in":"header","name":"X-Pi-Provider","required":false,"schema":{"type":"string"}},{"description":"Model id/pattern (default: server default, usually openrouter/free).","in":"header","name":"X-Pi-Model","required":false,"schema":{"type":"string"}},{"description":"Required for providers the server does not know. Known: openrouter, nvidia.","in":"header","name":"X-Pi-Base-Url","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"api_key":{"description":"Alternative to the X-Pi-Api-Key header.","type":"string"},"base_url":{"description":"Alternative to the X-Pi-Base-Url header.","type":"string"},"model":{"description":"Alternative to the X-Pi-Model header.","type":"string"},"provider":{"description":"Alternative to the X-Pi-Provider header.","type":"string"},"task":{"description":"The coding task in plain words. Be specific: language, what it should do, any constraints.","example":"Write a Python script that reads a CSV of sales data and prints total revenue per month.","maxLength":2000,"type":"string"}},"required":["task"],"type":"object"}}},"required":true},"responses":{"202":{"content":{"application/json":{"schema":{"properties":{"job_id":{"description":"Poll GET /api/pi-task/{job_id} for the result.","type":"string"},"status":{"enum":["queued"],"type":"string"}},"type":"object"}}},"description":"Task accepted and queued."},"400":{"description":"Missing or invalid task, or unknown provider without a base URL."},"401":{"description":"Missing X-Pi-Api-Key."},"429":{"description":"Rate limit exceeded."}},"security":[{"providerKey":[]}],"summary":"Ask Pi to do a coding task"}},"/api/pi-task/{job_id}":{"get":{"description":"Check a submitted task. status is queued, running, done, failed, or timeout. When done, result holds Pi's answer (usually code and/or an explanation). Your API key is never returned.","parameters":[{"description":"The job_id returned by POST /api/pi-task.","in":"path","name":"job_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"properties":{"created_at":{"type":"string"},"error":{"description":"Set when status is failed or timeout.","type":"string"},"finished_at":{"type":"string"},"id":{"type":"string"},"model":{"type":"string"},"provider":{"type":"string"},"result":{"description":"Pi's answer when status is done.","type":"string"},"status":{"enum":["queued","running","done","failed","timeout"],"type":"string"},"task":{"type":"string"}},"type":"object"}}},"description":"Job state."},"404":{"description":"Unknown job id."}},"summary":"Poll a Pi task for its result"}},"/health":{"get":{"description":"Liveness check.","responses":{"200":{"content":{"application/json":{"schema":{"properties":{"ok":{"type":"boolean"},"queued":{"type":"integer"},"running":{"type":"boolean"}},"type":"object"}}},"description":"Service is up."}},"summary":"Health check"}}},"servers":[{"url":"https://REPLACE-WITH-DEPLOYED-HOST"}]}